Don't run rubocop in CircleCI
For some reason, it's failing because it can't find certain gems. First, it complained about some gem I can't remember, and now it's complaining about panolint, which is not in the Gemfile.lock, and I don't have it locally either. I'm not sure why it's looking for it.

Also since we're in the CircleCI config, I updated the Postgres DB to match what's on Heroku now.