Update coffee-rails to 5.0
I had originally removed the coffee-rails gem because of this deprecation warning:

```
DEPRECATION WARNING: Single arity template handlers are deprecated. Template handlers must
now accept two parameters, the view object and the source for the view object.
Change:
  >> Coffee::Rails::TemplateHandler.call(template)
To:
  >> Coffee::Rails::TemplateHandler.call(template, source)
 (called from <top (required)> at /Users/moncef/projects/smcgov/ohana-api-smc/config/environment.rb:5)
```
But removing the gem cause asset precompilation to fail in CircleCI, so I added the gem back, but updated it to the latest 5.0 version, which got rid of the deprecation warning.