Update Ruby from 2.7.7 to 3.1.3
Ruby 2.7.7 will reach end of life at the end of March 2023, so now is a good time to update to 3.1.3, and then later to 3.2.1.

After changing the Ruby version in `.ruby-version` and running `bundle`, it failed to install `enumerable-statistic`, so I ran `bundle update enumerable-statistic`, and that fixed it. As I like to remind people, updating gems fixes 99% of issues. I also updated bundler for good measure.

Then I ran the tests, and they're all passing. In a separate commit, I'll fix the Rubocop offenses.