Prevent geocoding when not needed
It turns out that when deleting an organization from the admin interface, the address gets deleted first, which called the after_destroy callback, but in the Location model, the address wasn't being marked for destruction, so it attempted to geocode.

To fix this, we mark the locations for destruction in the controllers, and then we don't call this callback if the location is going to be destroyed.

Resetting coordinates is only when a location is going to stay, but its address is being removed.